Fast coded "Game Over" and "Win" states

+++ app/src/main/java/tk/erdmko/arcanoid/game/objects/BlockArray.java
@@ -59,7 +59,7 @@ public class BlockArray {
         }
     }
 
-    public void setSceneSize(int width, int height) {
+    public void setSceneSize(float width, float height) {
         this.sceneSize = new Vector2d(width, height);
         addObjects();
     }
